Shapes3

+++ kodilla-stream/src/main/java/com/kodilla/testing/shape/Shape.java
@@ -3,7 +3,7 @@ package com.kodilla.testing.shape;
 public interface Shape {
 
     public String getShapeName();
-    public int getField();
+    public double getField();
 
 
 }
